requirements.txt
requests
argparse
PyYAML